refactor: move teacher summary endpoint to a v2 controller
Route was registered in the legacy routes/entities/teachers dir,
which relies on routes/autohooks.ts global URL-prefix matching for
auth gating (PUBLIC_ROUTES list). New v2 endpoints don't go through
that hook — auth is opt-in per route via preHandler instead.

Extracted GET /entities/teachers/summary/:teacherId into
controllers/teacher-summary-controller.ts, registered directly on
the app (not under the /v2 prefix, to keep the existing path the
frontend depends on). No preHandler = explicitly public, decided at
the route definition rather than via the global PUBLIC_ROUTES list.

- Joabesv
