+++ apps/core/src/routes/entities/teachers/index.ts
@@ -3,7 +3,6 @@ import type { FastifyPluginAsyncZodOpenApi } from 'fastify-zod-openapi';
 import { Types } from 'mongoose';
 
 import { TeacherModel } from '@/models/Teacher.js';
-import { teacherSummarySchema } from '@/schemas/entities/summaries.js';
 import {
   createTeachersSchema,
   listTeachersSchema,
@@ -13,7 +12,6 @@ import {
 
 import {
   findAndUpdate,
-  findLatestSummary,
   findOne,
   listAll,
   populateWithSubject,
@@ -23,9 +21,6 @@ import {
 
 const plugin: FastifyPluginAsyncZodOpenApi = async (app) => {
   const teachersCache = app.cache<{}>();
-  const summaryCache = app.cache<
-    NonNullable<Awaited<ReturnType<typeof findLatestSummary>>>
-  >();
 
   app.get('/', { schema: listTeachersSchema }, async () => {
     const teachers = await listAll();
@@ -124,27 +119,6 @@ const plugin: FastifyPluginAsyncZodOpenApi = async (app) => {
 
     return resp;
   });
-
-  app.get(
-    '/summary/:teacherId',
-    { schema: teacherSummarySchema },
-    async (request, reply) => {
-      const { teacherId } = request.params;
-
-      const cacheKey = `summary:${teacherId}`;
-      const cached = summaryCache.get(cacheKey);
-      if (cached) return cached;
-
-      const summary = await findLatestSummary(teacherId);
-
-      if (!summary) {
-        return reply.notFound('Nenhum resumo disponível para esse professor');
-      }
-
-      summaryCache.set(cacheKey, summary);
-      return summary;
-    }
-  );
 };
 
 function getMean(value: any[], key?: string): any {
